Sending the pallete as a pointer

+++ Engine/src/engine.c
@@ -1343,7 +1343,7 @@ static void grouscan (int32_t dax1, int32_t dax2, int32_t sectnum, uint8_t  dast
             int32_t a3 = mulscale16(y2, zd) + (g_zx >> 6);
             slopevlin(ylookup[y2]+x+frameoffset,
                       krecip(a3>>3),
-                      (int32_t)nptr2,
+                      nptr2,
                       y2 - y1 + 1,
                       g_x1, g_y1,
                       a3,
